[refactor] 로직 부분 custom hook 분리
