+++ client/src/components/Ingredients/FormulaItem/index.tsx
@@ -1,22 +1,9 @@
 import React from 'react';
-import { LatexContent } from '../../../lib/constants/latex-header';
-import { useSelector } from 'react-redux';
-import { RootState } from '../../../contexts/index';
+import useFormulaItem, { FormulaItemProps } from './useFormulaItem';
 import * as S from './style';
-interface FormulaItemProps {
-  latexInfo: LatexContent;
-  hiddenFormula: () => void;
-}
 
 function FormulaItem({ latexInfo, hiddenFormula }: FormulaItemProps) {
-  const { mathfield } = useSelector((state: RootState) => state.latex);
-
-  const onClick = () => {
-    if (mathfield) {
-      mathfield.write(latexInfo.latex);
-      hiddenFormula();
-    }
-  };
+  const { onClick } = useFormulaItem({ latexInfo, hiddenFormula });
 
   return <S.FormulaItem onClick={onClick} image={latexInfo.image}></S.FormulaItem>;
 }
