+++ client/src/components/Ingredients/TextAreaItem/index.tsx
@@ -1,17 +1,9 @@
 import React from 'react';
-import { useDispatch, useSelector } from 'react-redux';
 import { TextArea } from './style';
-import { RootState } from '../../../contexts';
-import { editLatex } from '../../../contexts/latex';
+import useTextAreaItem from './useTextAreaItem';
 
 function TextAreaItem() {
-  const dispatch = useDispatch();
-  const { currentTab, totalLatex } = useSelector((state: RootState) => state.latex);
-  const currentText = totalLatex.filter((tabinfo) => tabinfo.id === currentTab)[0].latex;
-
-  const onChangeHandler = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    dispatch(editLatex(e.target.value));
-  };
+  const { currentText, onChangeHandler } = useTextAreaItem();
 
   return <TextArea value={currentText} onChange={onChangeHandler} />;
 }
