+++ client/src/components/Meal/Tab/MenuItem.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { useDispatch } from 'react-redux';
 import { Label, Menu } from 'semantic-ui-react';
 import { changeTab, removeTab, TabInfo } from '../../../contexts/latex';
+import useMenuItem from './useMunuItem';
 
 interface MenuItemProps {
   item: TabInfo;
@@ -10,20 +11,7 @@ interface MenuItemProps {
 }
 
 function MenuItem({ item, currentTab, index }: MenuItemProps) {
-  const dispatch = useDispatch();
-
-  const handleItemClick = (id: number) => {
-    dispatch(changeTab(id));
-  };
-
-  const removeTabHandler = (id: number) => {
-    dispatch(removeTab(id));
-  };
-
-  const onRemoveTab = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    removeTabHandler(item.id);
-  };
+  const { handleItemClick, onRemoveTab } = useMenuItem({ item });
 
   return (
     <Menu.Item
